+++ oc-includes/osclass/classes/utility/Escape.php
@@ -31,7 +31,8 @@ class Escape
         $str = strip_tags($str, $sNewLines);
         $str = str_replace("\r", '', $str);
         $str = addslashes($str);
-        $str = str_replace(array("\n", $aNewLines), '\n', $str);
+        $str = str_replace("\n", '\n', $str);
+        $str = str_replace($aNewLines, '\n', $str);
 
         return $str;
     }